refactor(cli): split up.rs into submodules and extract lifecycle helpers
Extract duplicated lifecycle code from the 2600-line up.rs:
- resolve_phase_entries(): replaces 2 identical phase-matching blocks
- run_config_phase_with_output(): replaces 3 identical progress+exec blocks
- Replace 3 inline LifecycleContext constructions with build_lifecycle_ctx()

Split into up/mod.rs (orchestration) and up/lifecycle.rs (phase resolution,
execution, content tracking, WaitForPhase).
